+++ src/systems/include/mp-units/systems/isq/mechanics.h
@@ -100,7 +100,7 @@ QUANTITY_SPEC(mechanical_work, energy, force* displacement, quantity_tensor_orde
 inline constexpr auto work = mechanical_work;
 QUANTITY_SPEC(mechanical_energy, mechanical_work, mass* pow<2>(length) / pow<2>(time));  // differs from ISO 80000
 QUANTITY_SPEC(potential_energy, mechanical_energy);                                      // differs from ISO 80000
-QUANTITY_SPEC(kinetic_energy, mechanical_energy, mass* pow<2>(speed));                   // differs from ISO 80000
+QUANTITY_SPEC(kinetic_energy, mechanical_energy, mass* pow<2>(speed), non_negative);     // differs from ISO 80000
 QUANTITY_SPEC(mechanical_efficiency, mechanical_power / mechanical_power, non_negative);
 QUANTITY_SPEC(mass_flow, mass_density* velocity);  // vector
 QUANTITY_SPEC(mass_flow_rate, mass_flow* area, quantity_tensor_order::scalar);